Restructure types, include proptypes as part of it
