--- src/typedef/Externals.jsdoc
@@ -1,48 +1,0 @@
-/**
- * Immutable.js Iterable
- * @typedef Iterable
- * @noexpand
- * @see https://facebook.github.io/immutable-js/docs/#/Iterable
- */
-
-/**
- * Immutable.js List
- * @typedef List
- * @noexpand
- * @see https://facebook.github.io/immutable-js/docs/#/List
- */
-
-/**
- * Immutable.js Map
- * @typedef Map
- * @noexpand
- * @see https://facebook.github.io/immutable-js/docs/#/Map
- */
-
-/**
- * Immutable.js OrderedMap
- * @typedef OrderedMap
- * @noexpand
- * @see https://facebook.github.io/immutable-js/docs/#/OrderedMap
- */
-
-/**
- * Immutable.js Record
- * @typedef Record
- * @noexpand
- * @see https://facebook.github.io/immutable-js/docs/#/Record
- */
-
-/**
- * React Component
- * @typedef ReactComponent
- * @noexpand
- * @see https://facebook.github.io/react/docs/react-component.html
- */
-
-/**
- * React Element
- * @typedef ReactElement
- * @noexpand
- * @see https://facebook.github.io/react/docs/react-component.html
- */
